refactor(reborn): tighten auth continuation composition
