+++ crates/ironclaw_reborn_composition/src/input.rs
@@ -5,7 +5,7 @@ use ironclaw_host_api::runtime_policy::EffectiveRuntimePolicy;
 use ironclaw_host_runtime::SchedulerTurnRunWakeNotifier;
 use ironclaw_trust::HostTrustPolicy;
 
-use crate::{RebornCompositionProfile, RebornProductAuthServices};
+use crate::{RebornCompositionProfile, RebornProductAuthServicePorts};
 
 pub struct RebornBuildInput {
     pub(crate) profile: RebornCompositionProfile,
@@ -17,7 +17,7 @@ pub struct RebornBuildInput {
     pub(crate) required_runtime_backends: Vec<ironclaw_host_api::RuntimeKind>,
     pub(crate) require_runtime_http_egress: bool,
     pub(crate) require_wasm_credentials: bool,
-    pub(crate) product_auth_services: Option<Arc<RebornProductAuthServices>>,
+    pub(crate) product_auth_ports: Option<RebornProductAuthServicePorts>,
 }
 
 pub(crate) enum RebornStorageInput {
@@ -163,13 +163,14 @@ impl RebornBuildInput {
         self
     }
 
-    /// Inject a Reborn-native product-auth composition bundle.
+    /// Inject Reborn-native product-auth service ports.
     ///
-    /// Production callers should provide durable implementations here once the
-    /// auth-flow and credential-account storage substrate is available. The
-    /// composition root never falls back to V1 route state or V1 secret stores.
-    pub fn with_product_auth_services(mut self, services: Arc<RebornProductAuthServices>) -> Self {
-        self.product_auth_services = Some(services);
+    /// Production callers should provide durable implementations here. The
+    /// composition root attaches the turn-continuation dispatcher after it has
+    /// composed the profile's [`ironclaw_turns::TurnCoordinator`], so OAuth
+    /// continuations cannot accidentally bypass the active coordinator.
+    pub fn with_product_auth_ports(mut self, ports: RebornProductAuthServicePorts) -> Self {
+        self.product_auth_ports = Some(ports);
         self
     }
 
@@ -188,7 +189,7 @@ impl RebornBuildInput {
             required_runtime_backends: Vec::new(),
             require_runtime_http_egress: false,
             require_wasm_credentials: false,
-            product_auth_services: None,
+            product_auth_ports: None,
         }
     }
 }
@@ -202,22 +203,14 @@ mod tests {
     use super::*;
 
     #[test]
-    fn with_product_auth_services_records_injected_bundle() {
-        let product_auth = Arc::new(
-            RebornProductAuthServices::from_shared_with_noop_dispatcher_for_tests(Arc::new(
-                InMemoryAuthProductServices::new(),
-            )),
-        );
-
-        let input = RebornBuildInput::disabled("test-owner")
-            .with_product_auth_services(Arc::clone(&product_auth));
-
-        assert!(Arc::ptr_eq(
-            input
-                .product_auth_services
-                .as_ref()
-                .expect("builder should retain injected product auth services"),
-            &product_auth
+    fn with_product_auth_ports_records_injected_ports() {
+        let product_auth = RebornProductAuthServicePorts::from_shared(Arc::new(
+            InMemoryAuthProductServices::new(),
         ));
+
+        let input =
+            RebornBuildInput::disabled("test-owner").with_product_auth_ports(product_auth.clone());
+
+        assert!(input.product_auth_ports.is_some());
     }
 }
