+++ crates/ironclaw_reborn_composition/src/lib.rs
@@ -46,7 +46,8 @@ use ironclaw_runtime_policy::{EffectiveRuntimePolicy as ResolvedRuntimePolicy, R
 
 pub use auth::{
     RebornAuthContinuationDispatcher, RebornOAuthCallbackError, RebornOAuthCallbackOutcome,
-    RebornOAuthCallbackRequest, RebornOAuthCallbackResponse, RebornProductAuthServices,
+    RebornOAuthCallbackRequest, RebornOAuthCallbackResponse, RebornProductAuthServicePorts,
+    RebornProductAuthServices,
 };
 pub use error::RebornBuildError;
 pub use factory::{RebornServices, build_reborn_services};
